refactor: function for sort items by index name
