+++ lib/api/links/link-data.ts
@@ -1,4 +1,5 @@
 import prisma from "@/lib/prisma";
+import { sortItemsByIndexAndName } from "@/lib/utils/sort-items-by-index-name";
 
 export async function fetchDataroomLinkData({ linkId }: { linkId: string }) {
   const linkData = await prisma.link.findUnique({
@@ -60,32 +61,10 @@ export async function fetchDataroomLinkData({ linkId }: { linkId: string }) {
     throw new Error("Dataroom not found");
   }
 
-  // Helper function to extract the numerical part of a string
-  const getNumber = (str: string): number => {
-    const match = str.match(/^\d+/);
-    return match ? parseInt(match[0], 10) : 0;
-  };
-
-  // Sort documents by orderIndex and/or name considering the numerical part
-  linkData.dataroom.documents.sort((a, b) => {
-    // First, compare by orderIndex if both items have it
-    if (a.orderIndex && b.orderIndex) {
-      if (a.orderIndex !== b.orderIndex) {
-        return a.orderIndex - b.orderIndex;
-      }
-    }
-
-    // If orderIndex is not available or equal, use name-based sorting
-    const numA = getNumber(a.document.name);
-    const numB = getNumber(b.document.name);
-
-    if (numA !== numB) {
-      return numA - numB;
-    }
-
-    // If numerical parts are the same, fall back to lexicographical order
-    return a.document.name.localeCompare(b.document.name);
-  });
+  // Sort documents by index or name
+  linkData.dataroom.documents = sortItemsByIndexAndName(
+    linkData.dataroom.documents,
+  );
 
   const brand = await prisma.dataroomBrand.findFirst({
     where: {
